Conflicts:
	bin/.gitignore
	bin/OBLFX/SubscriberCardController.class

+++ src/Server/CatalogQueries.java
@@ -11,7 +11,10 @@ import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.file.FileAlreadyExistsException;
+import java.nio.file.Files;
 import java.sql.Connection;
+import java.sql.ResultSet;
 
 import SystemObjects.ServerData;
 import SystemObjects.GeneralData.operationsReturn;
@@ -29,7 +32,7 @@ public class CatalogQueries {
 	 * enum for which sql query type using
 	 */
 	enum queryType {
-		Insert, Update
+	Insert, Update
 	};
 
 	static Connection con = mysqlConnection.conn;
@@ -49,6 +52,9 @@ public class CatalogQueries {
 				"INSERT INTO `obl`.`book` (`CatalogNumber`, `BookName`, `AuthorName`, `Subject`, `NumberOfCopies`, `AvailableCopies`, `NumberOfOrders`, `ShelfLocation`, `EditionNumber`, `purchesDate`, `isWanted`, `Description`, `ContextTable`)"
 						+ "VALUES (%s);",
 				bookDetailsQuery(bookToAdd, queryType.Insert));
+		String queryCopies = String.format("INSERT INTO `obl`.`bookcopy` (`CatalogNumber`) VALUES('%s');",
+				bookToAdd.getCatalogNumber());
+
 		// System.out.println(query);
 		try {
 			stmt = con.createStatement();
@@ -57,15 +63,32 @@ public class CatalogQueries {
 			FileOutputStream fos; // input from file
 			BufferedOutputStream bos; // buffer input
 			pdfoutFile = new File("src/PDFBook/" + bookToAdd.getContextTable());
+			if (pdfoutFile.exists())
+				throw new FileAlreadyExistsException("pdf file with this name already exist");
 			fos = new FileOutputStream(pdfoutFile);
 			bos = new BufferedOutputStream(fos);
 			bos.write(bookToAdd.getContextTableByteArray(), 0, bookToAdd.getContextTableByteArray().length); // read from file to byte array
 			bos.flush();
 			fos.flush();
 			fos.close();
-			result = new ServerData(operationsReturn.returnSuccessMsg, successMsg);
+
+			for (int i = 0; i < bookToAdd.getNumberOfLibraryCopies(); i++)
+				stmt.executeUpdate(queryCopies);
+
+			String queryCopiesAdded = String.format("SELECT * FROM obl.bookcopy where CatalogNumber='%s';",
+					bookToAdd.getCatalogNumber());
+			ResultSet copies = stmt.executeQuery(queryCopiesAdded);
+			String copiesIDs = "Copies ID added : ";
+			while (copies.next())
+				copiesIDs += String.format("%s, ", copies.getString("CopyID"));
+
+			copiesIDs = copiesIDs.substring(0, copiesIDs.length() - 2);
+
+			result = new ServerData(operationsReturn.returnSuccessMsg, successMsg + "\n" + copiesIDs);
 		} catch (MySQLIntegrityConstraintViolationException e) {
-			result = new ServerData(operationsReturn.returnException, new Exception("Catalog number already exist"));
+			result = new ServerData(operationsReturn.returnError, "Catalog number already exist");
+		} catch (FileAlreadyExistsException e) {
+			result = new ServerData(operationsReturn.returnError, e.getMessage());
 		} catch (SQLException | IOException e) {
 			result = new ServerData(operationsReturn.returnException, e);
 		}
@@ -107,9 +130,10 @@ public class CatalogQueries {
 		String query = "";
 		switch (queryType) {
 		case Insert:
-			query = String.format("'%s', '%s', '%s', '%s', 0, 0, 0, '%s', '%s', '%s', '%s', '%s', '%s'",
+			query = String.format("'%s', '%s', '%s', '%s', %d, %d, 0, '%s', '%s', '%s', '%s', '%s', '%s'",
 					bookWithDetails.getCatalogNumber(), bookWithDetails.getBookName(), bookWithDetails.getAuthorName(),
-					bookWithDetails.getSubject(), bookWithDetails.getShelfLoaction(),
+					bookWithDetails.getSubject(), bookWithDetails.getNumberOfLibraryCopies(),
+					bookWithDetails.getAvailableCopies(), bookWithDetails.getShelfLoaction(),
 					bookWithDetails.getEditionNumber(), dateFormat.format(bookWithDetails.getPurchesDate()),
 					bookWithDetails.getIsWanted() ? 1 : 0, bookWithDetails.getDescription(),
 					bookWithDetails.getContextTable());
